MOC (eliminación de archivos basura, arreglo del gtiignore)
