+++ .gitignore
@@ -6,6 +6,4 @@ node_modules/
 /blob-report/
 /playwright/.cache/
 /playwright/.auth/
-package.json
-package-lock.json
 .env
