+++ package.json
@@ -1,0 +1,17 @@
+{
+  "name": "proyectoleclere",
+  "version": "1.0.0",
+  "description": "este es el proyecto final del modulo 6 automatizar UI TRELLO",
+  "main": "index.js",
+  "scripts": {},
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@playwright/test": "^1.55.1",
+    "@types/node": "^24.5.2"
+  },
+  "dependencies": {
+    "dotenv": "^17.2.2"
+  }
+}
